A vast improvement
Made a searchbar for booking a flight, on index and trips. improved a lot of code. made dashboard print out more. a ton of functions added/improved. Register page DONE login page also done with the logic of it. Made a standard search. And lastly added the database as file